dramaForge_bac/requirements.txt
# DramaForge v2.0 Backend Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database & ORM
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.27.0
openai>=1.30.0

# File Handling
python-multipart>=0.0.9
python-docx>=1.1.0
aiofiles>=24.0.0

# Task Queue
arq>=0.26.0
redis>=5.0.0

# Media Processing
Pillow>=10.0.0

# Logging
loguru>=0.7.0
